Start to remove dupe code for blockchain nodes
