+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod bitcoin;
 pub mod create_comit_app;
-pub mod ethereum;
+pub mod docker;
 pub mod executable;
 pub mod start_env;
